old/requirements.txt
requests
beautifulsoup4
pandas
playwright
python-dotenv
openai

